Wobbly: Avoid freeing the VapourSynth nodes too early
Freeing a node before all frame requests from it are done will result
in a crash.

This could happen when opening a project which was saved when the crop
assistant was visible. When restoring the state of the application as
stored in the project, the crop assistant's visibility changes twice
somehow. This triggers two calls to evaluateMainDisplayScript in too
quick succession. The second call deletes the node created by the first
call before the frames requested by the first call are done.

Bug introduced in v3.