Wobbly: Fix bad frame when opening project with cropping
When opening a project which was saved when the crop assistant was
visible, the first frame displayed doesn't reflect the cropping
parameters.

What happens is that when restoring the application's state as
stored in the project, the crop assistant's visibility changes twice
somehow. This triggers two calls to evaluateMainDisplayScript in quick
succession. The first call finds the crop assistant not visible, and
therefore doesn't include CropRel and AddBorders in the script. The
second call happens before the frame requests from the first call are
done, so no frames are requested and the incorrect frame is left on
screen.

Some conditions for the bug introduced in v1, others in v3, but this bug
was never actually visible because of the crash fixed in the previous
commit.