Wobbly: Make the Preview tab update automatically
Most of the actions that affect the final script will now update the
preview immediately if the Preview tab is active. For example:
- changing the current frame's match
- adding or removing a freezeframe
- adding a range of frames to a custom list
- changing a section's presets
- etc.

This was requested in https://github.com/dubhater/Wobbly/issues/2.

One action that doesn't do this is editing a preset. It's too easy to
accidentally create an unusable preset, which would make an error
message appear. So when editing a preset switch to the Source tab and
back to the Preview tab to make it update.