Wobbly: Fix opening projects with bad cropping
Bad cropping or other problems that make it generate unusable scripts.

It is possible to save a project with invalid cropping, e.g. odd
numbers incompatible with the video's subsampling. When opening such
a project, script evaluation fails, but that doesn't have to leave the
application in a semi-invalid state. Instead of closing the previously
opened project and refusing to open the new one, open the new one in the
hopes that the user can fix the problem. In the case of bad cropping,
the user can fix it easily.